Change comments TO DO -> TODO. Add automatic recognition of drawers. Matrix as an exposed class.
