+++ src/webgldrawer.js
@@ -35,92 +35,6 @@
 
 (function( $ ){
 
-    // internal class Mat3: implements matrix operations
-    // Modified from https://webglfundamentals.org/webgl/lessons/webgl-2d-matrices.html
-   class Mat3{
-       constructor(values){
-           if(!values) {
-               values = [
-                   0, 0, 0,
-                   0, 0, 0,
-                   0, 0, 0
-               ];
-           }
-
-           this.values = values;
-       }
-
-       static makeIdentity(){
-           return new Mat3([
-               1, 0, 0,
-               0, 1, 0,
-               0, 0, 1
-           ]);
-       }
-
-       static makeTranslation(tx, ty) {
-           return new Mat3([
-               1, 0, 0,
-               0, 1, 0,
-               tx, ty, 1,
-           ]);
-       }
-
-       static makeRotation(angleInRadians) {
-           var c = Math.cos(angleInRadians);
-           var s = Math.sin(angleInRadians);
-           return new Mat3([
-               c, -s, 0,
-               s, c, 0,
-               0, 0, 1,
-           ]);
-       }
-
-       static makeScaling(sx, sy) {
-           return new Mat3([
-               sx, 0, 0,
-               0, sy, 0,
-               0, 0, 1,
-           ]);
-       }
-
-       multiply(other) {
-           let a = this.values;
-           let b = other.values;
-
-           var a00 = a[0 * 3 + 0];
-           var a01 = a[0 * 3 + 1];
-           var a02 = a[0 * 3 + 2];
-           var a10 = a[1 * 3 + 0];
-           var a11 = a[1 * 3 + 1];
-           var a12 = a[1 * 3 + 2];
-           var a20 = a[2 * 3 + 0];
-           var a21 = a[2 * 3 + 1];
-           var a22 = a[2 * 3 + 2];
-           var b00 = b[0 * 3 + 0];
-           var b01 = b[0 * 3 + 1];
-           var b02 = b[0 * 3 + 2];
-           var b10 = b[1 * 3 + 0];
-           var b11 = b[1 * 3 + 1];
-           var b12 = b[1 * 3 + 2];
-           var b20 = b[2 * 3 + 0];
-           var b21 = b[2 * 3 + 1];
-           var b22 = b[2 * 3 + 2];
-           return new Mat3([
-               b00 * a00 + b01 * a10 + b02 * a20,
-               b00 * a01 + b01 * a11 + b02 * a21,
-               b00 * a02 + b01 * a12 + b02 * a22,
-               b10 * a00 + b11 * a10 + b12 * a20,
-               b10 * a01 + b11 * a11 + b12 * a21,
-               b10 * a02 + b11 * a12 + b12 * a22,
-               b20 * a00 + b21 * a10 + b22 * a20,
-               b20 * a01 + b21 * a11 + b22 * a21,
-               b20 * a02 + b21 * a12 + b22 * a22,
-           ]);
-       }
-
-   }
-
    /**
     * @class WebGLDrawer
     * @memberof OpenSeadragon
@@ -198,7 +112,7 @@
             // Delete all our created resources
             gl.deleteBuffer(this._secondPass.bufferOutputPosition);
             gl.deleteFramebuffer(this._glFrameBuffer);
-            // TO DO: if/when render buffers or frame buffers are used, release them:
+            // TODO: if/when render buffers or frame buffers are used, release them:
             // gl.deleteRenderbuffer(someRenderbuffer);
             // gl.deleteFramebuffer(someFramebuffer);
 
@@ -277,9 +191,9 @@
 
             let flipMultiplier = this.viewport.flipped ? -1 : 1;
             // calculate view matrix for viewer
-            let posMatrix = Mat3.makeTranslation(-viewport.center.x, -viewport.center.y);
-            let scaleMatrix = Mat3.makeScaling(2 / viewport.bounds.width * flipMultiplier, -2 / viewport.bounds.height);
-            let rotMatrix = Mat3.makeRotation(-viewport.rotation);
+            let posMatrix = $.Mat3.makeTranslation(-viewport.center.x, -viewport.center.y);
+            let scaleMatrix = $.Mat3.makeScaling(2 / viewport.bounds.width * flipMultiplier, -2 / viewport.bounds.height);
+            let rotMatrix = $.Mat3.makeRotation(-viewport.rotation);
             let viewMatrix = scaleMatrix.multiply(rotMatrix).multiply(posMatrix);
 
             gl.bindFramebuffer(gl.FRAMEBUFFER, null);
@@ -300,7 +214,7 @@
                                         tiledImage._croppingPolygons ||
                                         tiledImage.debugMode
                                     );
-                let useTwoPassRendering = useContext2dPipeline || (tiledImage.opacity < 1); // TO DO: check hasTransparency in addition to opacity
+                let useTwoPassRendering = useContext2dPipeline || (tiledImage.opacity < 1); // TODO: check hasTransparency in addition to opacity
 
 
                 let tilesToDraw = tiledImage.getTilesToDraw();
@@ -343,10 +257,10 @@
                 let imageRotation = tiledImage.getRotation(true);
                 // if needed, handle the tiledImage being rotated
                 if( imageRotation % 360 !== 0){
-                    let imageRotationMatrix = Mat3.makeRotation(-imageRotation * Math.PI / 180);
+                    let imageRotationMatrix = $.Mat3.makeRotation(-imageRotation * Math.PI / 180);
                     let imageCenter = tiledImage.getBoundsNoRotate(true).getCenter();
-                    let t1 = Mat3.makeTranslation(imageCenter.x, imageCenter.y);
-                    let t2 = Mat3.makeTranslation(-imageCenter.x, -imageCenter.y);
+                    let t1 = $.Mat3.makeTranslation(imageCenter.x, imageCenter.y);
+                    let t2 = $.Mat3.makeTranslation(-imageCenter.x, -imageCenter.y);
 
                     // update the view matrix to account for this image's rotation
                     let localMatrix = t1.multiply(imageRotationMatrix).multiply(t2);
@@ -431,7 +345,7 @@
                     // Draw the quad (two triangles)
                     gl.drawArrays(gl.TRIANGLES, 0, 6);
 
-                    // TO DO: is this the mechanism we want to use here?
+                    // TODO: is this the mechanism we want to use here?
                     // iterate over any filters - filters can use this._renderToTexture to get rendered data if desired
                     let filters = this.filters || [];
                     for(let fi = 0; fi < filters.length; fi++){
@@ -456,7 +370,7 @@
                 }
 
                 // Fire tiled-image-drawn event.
-                // TO DO: the image data may not be on the output canvas yet!!
+                // TODO: the image data may not be on the output canvas yet!!
                 if( this.viewer ){
                     /**
                         * Raised when a tiled image is drawn to the canvas. Only valid
@@ -477,7 +391,7 @@
                 }
 
             });
-            // TO DO: the line below is a test!
+            // TODO: the line below is a test!
             if(renderingBufferHasImageData){
                 this._outputContext.drawImage(this._renderingCanvas, 0, 0);
             }
@@ -572,7 +486,7 @@
             let w = right - x;
             let h = bottom - y;
 
-            let matrix = new Mat3([
+            let matrix = new $.Mat3([
                 w, 0, 0,
                 0, h, 0,
                 x, y, 1,
@@ -580,11 +494,11 @@
 
             if(tile.flipped){
                 // flip the tile around the center of the unit quad
-                let t1 = Mat3.makeTranslation(0.5, 0);
-                let t2 = Mat3.makeTranslation(-0.5, 0);
+                let t1 = $.Mat3.makeTranslation(0.5, 0);
+                let t2 = $.Mat3.makeTranslation(-0.5, 0);
 
                 // update the view matrix to account for this image's rotation
-                let localMatrix = t1.multiply(Mat3.makeScaling(-1, 1)).multiply(t2);
+                let localMatrix = t1.multiply($.Mat3.makeScaling(-1, 1)).multiply(t2);
                 matrix = matrix.multiply(localMatrix);
             }
 
@@ -596,7 +510,7 @@
 
             if(this.continuousTileRefresh){
                 // Upload the image into the texture
-                // TO DO: test if this works appropriately
+                // TODO: test if this works appropriately
                 let tileContext = tile.getCanvasContext();
                 this._raiseTileDrawingEvent(tiledImage, this._outputContext, tile, tileContext);
                 this._uploadImageData(tileContext, tile, tiledImage);
@@ -798,7 +712,7 @@
             gl.enableVertexAttribArray(this._secondPass.aTexturePosition);
 
             // set the matrix that transforms the framebuffer to clip space
-            let matrix = Mat3.makeScaling(2, 2).multiply(Mat3.makeTranslation(-0.5, -0.5));
+            let matrix = $.Mat3.makeScaling(2, 2).multiply($.Mat3.makeTranslation(-0.5, -0.5));
             gl.uniformMatrix3fv(this._secondPass.uMatrix, false, matrix.values);
         }
 
@@ -957,8 +871,6 @@
             } catch (e){
                 $.console.error('Error uploading image data to WebGL', e);
             }
-
-
         }
 
         _imageUnloadedHandler(event){
@@ -977,7 +889,7 @@
             }
 
             // release the position buffer from the GPU
-            // TO DO: do this!
+            // TODO: do this!
         }
         // private
         // necessary for clip testing to pass (test uses spyOnce(drawer._setClip))
